# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Solar Generation API",
    "version": "1.0.0",
    "description": "Implemented endpoints: application liveness, OpenAPI, user/admin and device JWT exchange, and jurisdiction-scoped province lists. Active installations can submit append-only readings. Users can retrieve individual readings within their stored jurisdiction. Users can page and filter installation reading history within current jurisdiction."
  },
  "servers": [
    {
      "url": "/api/v1.0"
    }
  ],
  "paths": {
...
    "/grid-substations/{substationId}": {
      "get": {
        "operationId": "getGridSubstation",
        "summary": "Retrieve public grid substation details within the current user jurisdiction",
        "description": "Require current user authentication (installation tokens return 401), shared 120/minute User read limits and a public UUID v4 substationId. Invalid format returns 400. Resolve complete substation/district/province ancestry in a read snapshot before data or validators. National analysts/admins read nationally; provincial/district analysts within stored jurisdiction; otherwise 403. Missing substation or ancestry returns 404. Return only id, districtId and name. Strong ETag covers current principal and complete public representation; private, no-cache. Matching If-None-Match returns bodyless 304 after authentication, jurisdiction and rate limiting. No reliable metadata change timestamp exists, so omit Last-Modified; If-Modified-Since alone returns 200. Errors use no-store without validators. No child installation list, filters or unrelated endpoints added.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "substationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Strong/weak matching tag, quoted tag list, or * returns bodyless 304 after authorization. Takes precedence over If-Modified-Since."
          },
          {
            "name": "If-Modified-Since",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Ignored because no reliable substation metadata Last-Modified exists; alone returns 200."
          }
        ],
        "responses": {
          "200": {
            "description": "Public grid substation details",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GridSubstation"
                }
              }
            }
          },
          "304": {
            "description": "Authorized public substation unchanged; no body or Content-Type",
            "headers": {
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              },
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              }
            }
          },
          "400": {
            "description": "substationId must be a public UUID (INVALID_REQUEST).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, expired or installation token, or deleted/invalid user (UNAUTHORIZED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "WWW-Authenticate": {
                "$ref": "#/components/headers/BearerChallenge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Substation outside stored jurisdiction (FORBIDDEN).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "The substation is outside your permitted jurisdiction.",
                  "details": []
                }
              }
            }
          },
          "404": {
            "description": "Missing substation or ancestry (NOT_FOUND).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "Substation not found.",
                  "details": []
                }
              }
            }
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "description": "Shared user-read limit exceeded (RATE_LIMIT_EXCEEDED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Retry-After": {
                "$ref": "#/components/headers/LoginRetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected persistence failure (INTERNAL_SERVER_ERROR).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/districts/{districtId}/grid-substations": {
      "get": {
        "operationId": "listDistrictGridSubstations",
        "summary": "List grid substations belonging to an authorized district",
        "description": "Require current user authentication (installation tokens return 401), shared 120/minute User read limits and a UUID v4 districtId (invalid format 400). Resolve District and Province before substation queries/counts/validators. Missing district/ancestry returns 404. National analysts/admins access nationally; provincial analysts districts in their stored province; district analysts only their assigned district; otherwise 403. Query only URL district substations. Return the full collection without pagination, ordered by name/publicId ascending. No query parameters are accepted; supplied options return 400. Return public id/districtId/name items in count/items envelope; empty authorized districts return 200 with count=0 and empty items. Parent authorization and full list share one read snapshot; count is the list length and pagination fields are omitted. ETag covers current principal, district and complete response; private, no-cache. Matching If-None-Match returns bodyless 304 after access/rate checks. Omit Last-Modified; If-Modified-Since alone returns 200. Errors use no-store without validators/data/counts.",
        "security": [
          {
            "UserBearer": []
          }
        ],
        "parameters": [
          {
            "name": "districtId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
            }
          },
          {
            "name": "If-None-Match",
            "in": "header",
            "schema": {
              "type": "string"
            },
            "description": "Matching strong/weak tag, tag in a list, or * returns bodyless 304 only after authorization and read limiting."
          }
        ],
        "responses": {
          "200": {
            "description": "Complete district-scoped substation collection with count/items only; no pagination fields",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              },
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GridSubstationList"
                }
              }
            }
          },
          "304": {
            "description": "Complete authorized response unchanged; no body or Content-Type.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/PrivateCacheControl"
              },
              "ETag": {
                "$ref": "#/components/headers/ScopedETag"
              }
            }
          },
          "400": {
            "description": "Invalid district UUID (INVALID_REQUEST) or unsupported query parameters (INVALID_QUERY).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Missing/invalid/expired token, installation token, or deleted/invalid user (UNAUTHORIZED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "WWW-Authenticate": {
                "$ref": "#/components/headers/BearerChallenge"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "District outside stored jurisdiction (FORBIDDEN), before substation queries.",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "FORBIDDEN",
                  "message": "The district is outside your permitted jurisdiction.",
                  "details": []
                }
              }
            }
          },
          "404": {
            "description": "Missing district or province ancestry (NOT_FOUND).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                },
                "example": {
                  "code": "NOT_FOUND",
                  "message": "District not found.",
                  "details": []
                }
              }
            }
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "description": "Shared user-read limit exceeded (RATE_LIMIT_EXCEEDED).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              },
              "Retry-After": {
                "$ref": "#/components/headers/LoginRetryAfter"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected persistence failure (INTERNAL_SERVER_ERROR).",
            "headers": {
              "Cache-Control": {
                "$ref": "#/components/headers/NoStore"
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message",
          "details"
        ],
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {}
          }
        }
      },
...
      "GridSubstation": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "districtId",
          "name"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "districtId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-4[0-9a-f]{3}-[89ab][0-9a-f]{3}-[0-9a-f]{12}$"
          },
          "name": {
            "type": "string",
            "minLength": 1
          }
        }
      },
      "GridSubstationList": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "count",
          "items"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "minimum": 0,
            "description": "Number of substations returned for the authorized URL district."
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GridSubstation"
            }
          }
        }
      },
...
    },
    "securitySchemes": {
      "UserBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "HS256 user JWT from /auth/user-tokens. Each protected request reloads the current User role and jurisdiction."
      },
      "InstallationBearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Installation JWT from /auth/device-tokens; requires installation-write scope and current active installation ownership."
      }
    },
    "headers": {
      "NoStore": {
        "schema": {
          "type": "string",
          "enum": [
            "no-store"
          ]
        }
      },
      "BearerChallenge": {
        "schema": {
          "type": "string",
          "enum": [
            "Bearer"
          ]
        }
      },
      "LoginRetryAfter": {
        "schema": {
          "type": "integer",
          "minimum": 1
        },
        "description": "Seconds until the exceeded counter window expires."
      },
      "PrivateCacheControl": {
        "schema": {
          "type": "string",
          "enum": [
            "private, no-cache"
          ]
        }
      },
      "ScopedETag": {
        "schema": {
          "type": "string"
        },
        "description": "Strong validator of the exact scoped representation and current stored principal."
      },
...
    },
    "responses": {
...
      "NotAcceptable": {
        "description": "Accept excludes JSON; no body"
      },
...
    }
  }
}